refactor: move `Clouds` from root to Renderer folder and fix include paths
