+++ source/InjectHooksMain.cpp
@@ -107,7 +107,7 @@
 #include "Skidmarks.h"
 #include "CarCtrl.h"
 #include "TagManager.h"
-#include "Clouds.h"
+#include "Renderer/Clouds.h"
 #include "StuntJumpManager.h"
 #include "CustomCarPlateMgr.h"
 #include "WindModifiers.h"
