+++ source/app/app_game.cpp
@@ -6,7 +6,7 @@
 #include "LoadingScreen.h"
 #include "PlantMgr.h"
 #include "Shadows.h"
-#include "Clouds.h"
+#include "Renderer/Clouds.h"
 #include "MovingThings.h"
 #include "BreakManager_c.h"
 #include "RealTimeShadowManager.h"
@@ -16,7 +16,7 @@
 #include "Garages.h"
 #include "UIRenderer.h"
 #include "Gamma.h"
-#include "../game_sa/Renderer/Birds.h"
+#include "Renderer/Birds.h"
 #include "Skidmarks.h"
 #include "Ropes.h"
 #include "Glass.h"
