+++ source/game_sa/Renderer/Clouds.cpp
@@ -1,7 +1,7 @@
 #include "StdInc.h"
 
 #include "Clouds.h"
-#include "PostEffects.h"
+#include "../PostEffects.h"
 
 auto& CurrentFogIntensity = StaticRef<float>(0x8D5798);
