+++ source/game_sa/Renderer/Clouds.h
