+++ source/toolsmenu/DebugModules/CloudsDebugModule.cpp
@@ -1,6 +1,6 @@
 #include <StdInc.h>
 #include "CloudsDebugModule.hpp"
-#include <Clouds.h>
+#include <Renderer/Clouds.h>
 
 using namespace ImGui;
